requirements.txt
Flask==0.12
tweepy==3.5.0
requests-toolbelt==0.7.1
google-cloud-language=0.23.1
